+++ CHANGELOG.md
@@ -17,6 +17,27 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [example] Added axum example
 - [example] Added fibonacci sequence example
 
+## [2.1.2] - 2026.05.20
+
+### Fixed
+
+- [framework] `Body` `Read` impl hangs (https://github.com/wasselteam/sdk-rust/issues/1)
+
+### Added
+
+- [framework] `From<Bytes>` implementation for Body
+
+### Changed
+
+- [examples] Make examples members of workspace
+- [framework] Remove `Read` implementation from body
+
+## [2.1.1] - 2026.05.12
+
+### Fixed
+
+- [framework] Fix panic in Read impl of Body
+
 ## [2.1.0] - 2026.05.12
 
 ### Added